refactor : Book/serv Impl 제거
